# 8307348 - Parallelize heap walk for ObjectCount(AfterGC) JFR event collection