@olivergillespie 2023-05-04
Passed tier2 and tier3 tests for me locally on Linux x86_64 and aarch64, with G1, Serial and Parallel, with just a few (14) tier2 failures in `java/net/httpclient/` which I believe are due to a bad configuration in my local setup.