@tschatzl 2023-05-04
> I looked if there might be a better option, like passing the `WorkerThreads*` from callers to actually figure out the number of active workers from the GC itself, but all of this cuts rather deep. I would say that should be done in a separate PR. `ParallelGCThreads` should work well meanwhile.
> 
> @albertnetymk, @tschatzl might have an opinion here.

As you suggested, the typical way is passing the `WorkerThreads*` along instead of passing a thread number and the code selecting the `WorkerThread*` by itself. Actually I'm not sure why safepoint_workers are used at all.

The changes to do that does not seem to be that significant to me actually, so I would prefer that. The current change is an improvement already though.