@albertnetymk 2023-05-04
> Actually I'm not sure why safepoint_workers are used at all.

I believe it's semantically incorrect to use `safepoint_workers` here. Maybe `HeapInspection` should live in `gc` folder.

> The changes to do that does not seem to be that significant to me actually, so I would prefer that.

How does that affect another caller, `VM_GC_HeapInspection::doit`? Unclear to me how one can get gc-workers in that context.

Could one introduce another API in `class CollectedHeap`, sth like `virtual WorkerThreads* gc_workers() { return nullptr; }`, next to `safepoint_workers`?