@albertnetymk 2023-05-05
> I think safepoint_workers is "fine" to use from VM_GC_HeapInspection.

I see; the name contains "GC", but it doesn't really correspond to a gc-safepoint (young/full gc pause). (To make matters worse, that vm-op can potentially perform a full-gc...)

Then, `populate_table` will get diff workers, depending whether it's called inside a gc-safepoint or not.